requirements.txt
Django==1.9.8
django-bootstrap3==7.0.1
django-decorator-include==1.1
django-extensions==1.6.7
django-simple-menu==1.2.1
django-tables2==1.2.5
factory-boy==2.7.0
fake-factory==0.5.9
future==0.15.2
python-dateutil==2.5.3
six==1.10.0
Werkzeug==0.11.10
